Completely moved the regressor command into the train command as a subcommand. The classifier is next.
